requirements.txt
python-telegram-bot~=13.0
environs~=9.5.0
google_cloud_dialogflow~=2.21.0
vk-api~=11.9.9
tqdm~=4.65.0
